DataTableBuilder : introduction d'un toolbar builder (ajustements).
git-svn-id: svn+ssh://svn.projects.openwide.fr/svnroot/owsi-tools/branches/owsi-core-0.9-MAINTENANCE@4225 d3474844-eb8e-4abc-b058-2b321fed648b

+++ owsi-core/owsi-core-components/owsi-core-component-wicket-more/src/main/java/fr/openwide/core/wicket/more/markup/html/repeater/data/table/builder/toolbar/state/IAddedToolbarCoreElementState.java
@@ -12,7 +12,7 @@ public interface IAddedToolbarCoreElementState<T, S extends ISort<?>> extends IA
 	IAddedToolbarCoreElementState<T, S> withClass(String cssClass);
 
 	@Override
-	IAddedToolbarCoreElementState<T, S> colspan(long colspan);
+	IAddedToolbarCoreElementState<T, S> colspan(Integer colspan);
 
 	@Override
 	IAddedToolbarCoreElementState<T, S> full();
