+++ owsi-core/owsi-core-components/owsi-core-component-wicket-more/src/main/java/fr/openwide/core/wicket/more/markup/html/repeater/data/table/builder/toolbar/state/IAddedToolbarLabelElementState.java
@@ -12,7 +12,7 @@ public interface IAddedToolbarLabelElementState<T, S extends ISort<?>> extends I
 	IAddedToolbarLabelElementState<T, S> withClass(String cssClass);
 
 	@Override
-	IAddedToolbarLabelElementState<T, S> colspan(long colspan);
+	IAddedToolbarLabelElementState<T, S> colspan(Integer colspan);
 
 	@Override
 	IAddedToolbarLabelElementState<T, S> full();
